requirements.txt
numpy==2.1.3
torch==2.6.0
torchvision==0.21.0
Pillow==11.1.0
scikit-learn==1.6.1
matplotlib==3.10.1
Flask==3.1.0